tests: new tests and fix problems in tests
